pyproject.toml
[tool.poetry]
name = "cinema"
version = "0.1.0"
description = ""
authors = ["top1Dron <pro100moneyfarmer@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.3"
psycopg2 = "^2.8.6"
django-multiselectfield = "^0.1.12"
Pillow = "^8.2.0"
django-debug-toolbar = "^3.2.1"
django-widget-tweaks = "^1.4.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
